+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Providing a job list to `Batch.__init__()` is now optional; an empty batch will 
   be created if the job list is not provided
 - `Batch.__init__()` no longer issues a warning when creating an empty batch
+- `HyP3.find_jobs()` will now issue a warning when a zero jobs were found
 
 ## [0.3.2](https://github.com/ASFHyP3/hyp3-sdk/compare/v0.3.1...v0.3.2)
 ### Changed